+++ lib/features/analytics/mixpanels/analytics_repository.dart
@@ -1,10 +1,14 @@
+// ignore_for_file: depend_on_referenced_packages, constant_identifier_names
+
 import 'package:dartz/dartz.dart';
 import 'package:helixworlds_snatcher_sdk/core/failure.dart';
 import 'package:helixworlds_snatcher_sdk/core/success.dart';
 import 'package:helixworlds_snatcher_sdk/features/analytics/mixpanels/analytics_googleanalytics_remote_datasource.dart';
 import 'package:helixworlds_snatcher_sdk/features/analytics/mixpanels/analytics_mixpanels_remote_datasource.dart';
+import 'package:helixworlds_snatcher_sdk/features/analytics/mixpanels/analytics_mixpanels_rudderstack_remote_datasource.dart';
 import 'package:helixworlds_snatcher_sdk/features/scan/data/model/scan_model.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:rudder_sdk_flutter_platform_interface/platform.dart';
 
 abstract class IAnalyticsRepository {
 
@@ -17,12 +21,19 @@ abstract class IAnalyticsRepository {
   Future<Either<Failure, Success>> analyticsSavedItems(InventoryItemModel item);
 }
 const String localKeyInstallation = "localKeyInstallation";
+const String analytics_installKey = "installs";
+const String analytics_scannedItems = "scanned_items";
+const String analytics_redirectToShop = "redirect_to_shop";
+const String analytics_savedItems = "saved_items";
+const String analytics_attemptScannedItems = "attempt_scanned_items";
+const String analytics_redirectToShopEventId = "analytics_redirectToShopEventId";
 
 class AnalyticsRepository extends IAnalyticsRepository {
   final IAnalyticsMixpanelsRemoteDatasource? mixPanelRemoteDS;
   final SharedPreferences _sharedPref;
   final IGoogleAnalyticsRemoteDatasource? googleAnalyticsRemoteDS;
-  AnalyticsRepository(this._sharedPref, {this.googleAnalyticsRemoteDS, this.mixPanelRemoteDS});
+  final IAnalyticsRudderStackRemoteDatasource? rudderStackRemoteDS;
+  AnalyticsRepository(this._sharedPref, {this.googleAnalyticsRemoteDS, this.mixPanelRemoteDS, this.rudderStackRemoteDS});
 
   @override
   Future<Either<Failure, Success>> analyticsTrackInstalls() async {
@@ -30,10 +41,13 @@ class AnalyticsRepository extends IAnalyticsRepository {
       var result = _sharedPref.getBool(localKeyInstallation) ?? false;
       if(result == false){
         if(mixPanelRemoteDS != null){
-          await mixPanelRemoteDS?.trackEvent("installs", {});
+          await mixPanelRemoteDS?.trackEvent(analytics_installKey, {});
         }
         if(googleAnalyticsRemoteDS != null){
-          await googleAnalyticsRemoteDS?.sendAnalytics("installs", {});
+          await googleAnalyticsRemoteDS?.sendAnalytics(analytics_installKey, {});
+        }
+        if(rudderStackRemoteDS != null){
+          await rudderStackRemoteDS?.trackEvent(analytics_installKey, RudderProperty().putValue(key: analytics_installKey, value: {}));
         }
         return Right(AnalyticsLogsSuccess());
       } else {
@@ -49,10 +63,13 @@ class AnalyticsRepository extends IAnalyticsRepository {
   Future<Either<Failure, Success>> analyticsScannedItems(InventoryItemModel item) async {
     try {
         if(mixPanelRemoteDS != null){
-          await mixPanelRemoteDS?.trackEvent("scanned_items", item.toJson());
+          await mixPanelRemoteDS?.trackEvent(analytics_scannedItems, item.toJson());
         }
         if(googleAnalyticsRemoteDS != null){
-          await googleAnalyticsRemoteDS?.sendAnalytics("scanned_items", item.toJson());
+          await googleAnalyticsRemoteDS?.sendAnalytics(analytics_scannedItems, item.toJson());
+        }
+        if(rudderStackRemoteDS != null){
+          await rudderStackRemoteDS?.trackEvent(analytics_scannedItems, RudderProperty().putValue(key: analytics_scannedItems, value: item.toJson()));
         }
         return Right(AnalyticsLogsSuccess());
     } catch(e) {
@@ -63,11 +80,15 @@ class AnalyticsRepository extends IAnalyticsRepository {
   @override
   Future<Either<Failure, Success>> analyticsRedirectToShopEvent(String url, String userId) async {
     try {
+        dynamic value = {"url": url, "userId": userId};
         if(mixPanelRemoteDS != null){
-          await mixPanelRemoteDS?.trackEvent("redirect_to_shop", {"url": url, "userId": userId});
+          await mixPanelRemoteDS?.trackEvent(analytics_redirectToShop, value);
         }
         if(googleAnalyticsRemoteDS != null){
-          await googleAnalyticsRemoteDS?.sendAnalytics("redirect_to_shop", {"url": url, "userId": userId});
+          await googleAnalyticsRemoteDS?.sendAnalytics(analytics_redirectToShop, value);
+        }
+        if(rudderStackRemoteDS != null){
+          await rudderStackRemoteDS?.trackEvent(analytics_redirectToShop, RudderProperty().putValue(key: analytics_redirectToShop, value: value));
         }
         return Right(AnalyticsLogsSuccess());
     } catch(e) {
@@ -79,10 +100,13 @@ class AnalyticsRepository extends IAnalyticsRepository {
   Future<Either<Failure, Success>> analyticsSavedItems(InventoryItemModel item) async {
     try {
         if(mixPanelRemoteDS != null){
-          await mixPanelRemoteDS?.trackEvent("saved_items", item.toJson());
+          await mixPanelRemoteDS?.trackEvent(analytics_savedItems, item.toJson());
         }
         if(googleAnalyticsRemoteDS != null){
-          await googleAnalyticsRemoteDS?.sendAnalytics("saved_items", item.toJson());
+          await googleAnalyticsRemoteDS?.sendAnalytics(analytics_savedItems, item.toJson());
+        }
+        if(rudderStackRemoteDS != null){
+          await rudderStackRemoteDS?.trackEvent(analytics_savedItems, RudderProperty().putValue(key: analytics_savedItems, value: item.toJson()));
         }
         return Right(AnalyticsLogsSuccess());      
     } catch(e) {
@@ -93,11 +117,16 @@ class AnalyticsRepository extends IAnalyticsRepository {
   @override
   Future<Either<Failure, Success>> analyticsRedirectToShopEventItemId(String url, String userId, String itemId) async {
     try {
+        dynamic value = {"url": url, "userId": userId, "itemId": itemId };
+
         if(mixPanelRemoteDS != null){
-          await mixPanelRemoteDS?.trackEvent("redirect_to_shop", {"url": url, "userId": userId, "itemId": itemId });
+          await mixPanelRemoteDS?.trackEvent(analytics_redirectToShopEventId, value);
         }
         if(googleAnalyticsRemoteDS != null){
-          await googleAnalyticsRemoteDS?.sendAnalytics("redirect_to_shop", {"url": url, "userId": userId, "itemId": itemId });
+          await googleAnalyticsRemoteDS?.sendAnalytics(analytics_redirectToShopEventId, value);
+        }
+        if(rudderStackRemoteDS != null){
+          await rudderStackRemoteDS?.trackEvent(analytics_redirectToShopEventId, RudderProperty().putValue(key: analytics_savedItems, value: value));
         }
         return Right(AnalyticsLogsSuccess());      
     } catch(e) {
@@ -108,11 +137,15 @@ class AnalyticsRepository extends IAnalyticsRepository {
   @override
   Future<Either<Failure, Success>> analyticsAttemptScannedItems(Map<String, dynamic> scanResult, InventoryItemModel expectedItem) async {
     try {
+        dynamic value = {"scan_result": scanResult, "expected_item": expectedItem.toJson() };
         if(mixPanelRemoteDS != null){
-          await mixPanelRemoteDS?.trackEvent("attempt_scanned_items", {"scan_result": scanResult, "expected_item": expectedItem.toJson() });
+          await mixPanelRemoteDS?.trackEvent(analytics_attemptScannedItems, value);
         }
         if(googleAnalyticsRemoteDS != null){
-          await googleAnalyticsRemoteDS?.sendAnalytics("attempt_scanned_items", {"scan_result": scanResult, "expected_item": expectedItem.toJson() });
+          await googleAnalyticsRemoteDS?.sendAnalytics(analytics_attemptScannedItems, value);
+        }
+        if(rudderStackRemoteDS != null){
+          await rudderStackRemoteDS?.trackEvent(analytics_attemptScannedItems, RudderProperty().putValue(key: analytics_savedItems, value: value));
         }
         return Right(AnalyticsLogsSuccess());      
     } catch(e) {